增加评论的接口+异常处理

+++ src/main/java/life/majiang/community/community/controller/PublishController.java
@@ -21,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
+    public String edit(@PathVariable(name="id") Long id,
                        Model model){
         QuestionDTO questionDTO = questionService.listById(id);
         model.addAttribute("title", questionDTO.getTitle());
@@ -40,7 +40,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title") String title,
                             @RequestParam("description") String description,
                             @RequestParam("tag") String tag,
-                            @RequestParam(value = "id", defaultValue = "") Integer id,
+                            @RequestParam(value = "id", defaultValue = "") Long id,
                             HttpServletRequest request,
                             Model model){
